+++ shellfoundry/exceptions.py
@@ -36,7 +36,8 @@ class PlatformNameIsEmptyException(ShellFoundryBaseException):
 
 class FatalError(click.ClickException):
     def __init__(self, message):
-        super(click.ClickException, self).__init__(message)
+        # super(click.ClickException, self).__init__(message)
+        super(FatalError, self).__init__(message)
 
     def show(self, file=None):
         click.secho('Error: {}'.format(self.format_message()), err=True, fg='red')