+++ shellfoundry/utilities/cookiecutter_integration.py
@@ -4,7 +4,9 @@
 import datetime
 import os
 
+from click import ClickException
 from cookiecutter.main import cookiecutter
+from cookiecutter.exceptions import OutputDirExistsException
 from shellfoundry.utilities.config_reader import Configuration, CloudShellConfigReader
 from shellfoundry.utilities.constants import TEMPLATE_INFO_FILE
 
@@ -13,11 +15,17 @@ class CookiecutterTemplateCompiler(object):
     def __init__(self):
         self.cloudshell_config_reader = Configuration(CloudShellConfigReader())
 
-    def compile_template(self, shell_name, template_path, extra_context, running_on_same_folder):
+    def compile_template(self, shell_name, template_path, extra_context, running_on_same_folder, python_version=None):
+
+        if python_version is None:
+            python_version = ""
+        else:
+            python_version = ' PythonVersion="{}"'.format(str(python_version))
 
         extra_context.update({"project_name": shell_name,
                               "full_name": self.cloudshell_config_reader.read().author,
                               "release_date": datetime.datetime.now().strftime("%B %Y"),
+                              "python_version": str(python_version),
                               })
 
         if running_on_same_folder:
@@ -25,10 +33,12 @@ class CookiecutterTemplateCompiler(object):
         else:
             output_dir = os.path.curdir
 
-        cookiecutter(template_path, no_input=True,
-                     extra_context=extra_context,
-                     overwrite_if_exists=False, output_dir=output_dir)
-
+        try:
+            cookiecutter(template_path, no_input=True,
+                         extra_context=extra_context,
+                         overwrite_if_exists=False, output_dir=output_dir)
+        except OutputDirExistsException as err:
+            raise ClickException(str(err))
         # self._remove_template_info_file(output_dir)
 
     @staticmethod