+++ shellfoundry/utilities/template_versions.py
@@ -1,4 +1,6 @@
-import ast
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
+
 import requests
 import shellfoundry.exceptions as exc
 
@@ -20,18 +22,16 @@ class TemplateVersions(object):
         self.template_repo = [url_user, url_repo]
 
     def get_versions_of_template(self):
-        '''
+        """
         Get all versions (branches) of a given template.
         Raises HTTPError on request fail, NoVersionsHaveBeenFoundException when no versions have been found
         :return: List filled with version names (e.g. 1.0, 1.1, 2.0...)
-        '''
+        """
         response = requests.get(VERSIONS_URL.format(*self.template_repo))
-        if response.status_code != requests.codes.ok:
-            raise requests.HTTPError('Failed to receive versions from host')
+        response.raise_for_status()
 
-        response_arr = ast.literal_eval(response.text)
-        branches = [d[NAME_PLACEHOLDER] for d in response_arr]
-        branches.sort(reverse=True, key=lambda x:(is_version(x), x))
+        branches = [d[NAME_PLACEHOLDER] for d in response.json()]
+        branches.sort(reverse=True, key=lambda x: (is_version(x), x))
         if not self.has_versions(branches):
             raise exc.NoVersionsHaveBeenFoundException("No versions have been found for this template")
         return branches